Services/sentiment/pyproject.toml
[project]
name = "sentiment"
version = "0.1.0"
description = "Add your description here"
readme = "README.md"
requires-python = ">=3.12"
dependencies = [
    "fire>=0.7.0",
    "huggingface-hub>=0.27.0",
    "llama-cpp-python>=0.3.5",
    "llama-index-core>=0.12.5",
    "llama-index-llms-ollama>=0.5.0",
    "llama-index-llms-openai>=0.3.10",
    "loguru>=0.7.3",
    "pandas>=2.2.3",
    "pydantic>=2.10.3",
    "pydantic-settings>=2.6.1",
    "quixstreams>=3.4.0",
    "rarfile>=4.2",
    "requests>=2.32.3",
    "tqdm>=4.67.1",
]
